JENKINS-45922# Add indexing capability to MBP (#1292)
* JENKINS-45922# Add indexing capability to MBP

This is to allow pipeline editor to trigger indexing when it needs to.
This can also allow blueocean UI to allow triggering.

* JENKINS-45944# Added BluePipelineScm capability

This is going to fix pipeline editor's hardcoded comparison with _class
of pipeline and use BluePipelineScm capability.
