+++ blueocean-pipeline-api-impl/src/main/java/io/jenkins/blueocean/rest/impl/pipeline/OrganizationFolderRunContainerImpl.java
@@ -3,7 +3,6 @@ package io.jenkins.blueocean.rest.impl.pipeline;
 import com.google.common.collect.ImmutableList;
 import io.jenkins.blueocean.rest.Reachable;
 import io.jenkins.blueocean.rest.hal.Link;
-import io.jenkins.blueocean.rest.model.BlueOrganization;
 import io.jenkins.blueocean.rest.model.BlueRun;
 import io.jenkins.blueocean.rest.model.BlueRunContainer;
 import org.kohsuke.stapler.StaplerRequest;
