+++ blueocean-pipeline-api-impl/src/test/java/io/jenkins/blueocean/rest/impl/pipeline/MultiBranchTest.java
@@ -6,6 +6,7 @@ import hudson.Util;
 import hudson.model.FreeStyleProject;
 import hudson.model.Job;
 import hudson.model.Queue;
+import hudson.model.User;
 import hudson.plugins.favorite.Favorites;
 import hudson.plugins.git.util.BuildData;
 import hudson.scm.ChangeLogSet;
@@ -349,6 +350,27 @@ public class MultiBranchTest extends PipelineBaseTest {
             assertEquals(100, s);
         }
     }
+
+    @Test
+    public void multiBranchPipelineIndex() throws Exception {
+        User user = login();
+        WorkflowMultiBranchProject mp = j.jenkins.createProject(WorkflowMultiBranchProject.class, "p");
+        mp.getSourcesList().add(new BranchSource(new GitSCMSource(null, sampleRepo.toString(), "", "*", "", false),
+                new DefaultBranchPropertyStrategy(new BranchProperty[0])));
+        for (SCMSource source : mp.getSCMSources()) {
+            assertEquals(mp, source.getOwner());
+        }
+
+        Map map = new RequestBuilder(baseUrl)
+                .post("/organizations/jenkins/pipelines/p/runs/")
+                .jwtToken(getJwtToken(j.jenkins, user.getId(), user.getId()))
+                .data(ImmutableMap.of())
+                .status(200)
+                .build(Map.class);
+
+        assertNotNull(map);
+    }
+
 
     @Test
     public void getMultiBranchPipelineRuns() throws Exception {
@@ -833,13 +855,28 @@ public class MultiBranchTest extends PipelineBaseTest {
 
         j.waitUntilNoActivity();
 
-        Map response = get("/organizations/jenkins/pipelines/p/branches/master/", Map.class);
+        //check MBP capabilities
+        Map<String,Object> response = get("/organizations/jenkins/pipelines/p/");
         String clazz = (String) response.get("_class");
 
         response = get("/classes/"+clazz+"/");
         Assert.assertNotNull(response);
 
         List<String> classes = (List<String>) response.get("classes");
+        Assert.assertTrue(classes.contains(BLUE_SCM)
+                && classes.contains(JENKINS_MULTI_BRANCH_PROJECT)
+                && classes.contains(BLUE_MULTI_BRANCH_PIPELINE)
+                && classes.contains(BLUE_PIPELINE_FOLDER)
+                && classes.contains(JENKINS_ABSTRACT_FOLDER)
+                && classes.contains(BLUE_PIPELINE));
+
+        response = get("/organizations/jenkins/pipelines/p/branches/master/", Map.class);
+        clazz = (String) response.get("_class");
+
+        response = get("/classes/"+clazz+"/");
+        Assert.assertNotNull(response);
+
+        classes = (List<String>) response.get("classes");
         Assert.assertTrue(classes.contains(JENKINS_JOB)
             && classes.contains(JENKINS_WORKFLOW_JOB)
             && classes.contains(BLUE_BRANCH)
